tomd: do not try to create logical volumnes if they already exist -- just bail

+++ home/image_builders/grow.rb
@@ -13,9 +13,12 @@ define "grow" do
       cmd "losetup -c /dev/#{spec[:loop0]}"
       vm_disk_location = "/dev/#{spec[:loop0]}"
     when 'lvm'
+      vm_disk_location = "/dev/#{spec[:lvm_vg]}/#{spec[:hostname]}"
+      if File.exists?(vm_disk_location)
+        raise "Logical volume '#{vm_disk_location}' already exists!"
+      end
       cmd "lvcreate -n #{spec[:hostname]} -L #{spec[:image_size]} #{spec[:lvm_vg]}"
       cmd "dd if=#{spec[:images_dir]}/gold/generic.img of=/dev/#{spec[:lvm_vg]}/#{spec[:hostname]}"
-      vm_disk_location = "/dev/#{spec[:lvm_vg]}/#{spec[:hostname]}"
     else
       raise "provisioning tools does not know about vm_storage_type '#{config[:vm_storage_type]}'"
     end
